+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Command line switch to override config file name
 - Allows for only/except for specific cops by name or pattern 
 - Allows for excluding specific file patterns from being changed
+- Only runs rubocop on selected files, not on the whole repo
 
 ## [0.1.0] - 2022-07-20